requirements.txt
python-dotenv
openai
qdrant-client
pdf2image
pytesseract
langchain
langchain-openai
tqdm 